+++ output.h
@@ -7,5 +7,5 @@ void display(void *data);
 void displayFlags(Element *e, int flags);
 void blankBoard();
 void displayElements(Element *e);
-void emptyRectangle(uint x, uint y, uint width,
-			uint height, char display);
+void emptyRectangle(uint8 x, uint8 y, uint8 width,
+			uint8 height, char display);
