+++ room.h
@@ -3,7 +3,7 @@
 #define ROOM_H
 #include "structs.h" /* place the structs here instead of in this header */
 
-Room *makeRoom(uint internalWidth, uint internalHeight);
+Room *makeRoom(uint8 internalWidth, uint8 internalHeight);
 
 #endif
