+++ wall.h
@@ -4,6 +4,6 @@
 #include "structs.h"
 
 void displayWall(Element *e);
-Wall *makeWall(uint x, uint y, uint w, uint h, char vis);
+Wall *makeWall(uint8 x, uint8 y, uint8 w, uint8 h, char vis);
 int collideWall(Element *wall, Element *non);
 #endif
